refactor: finalisation de la nouvelle architecture
